+++ src/plugins/replication/index.ts
@@ -57,6 +57,7 @@ import {
 } from '../../rx-database-internal-store';
 import { addRxPlugin } from '../../plugin';
 import { hasEncryption } from '../../rx-storage-helper';
+import { overwritable } from '../../overwritable';
 
 
 export const REPLICATION_STATE_BY_COLLECTION: WeakMap<RxCollection, RxReplicationState<any, any>[]> = new WeakMap();
@@ -146,7 +147,8 @@ export class RxReplicationState<RxDocType, CheckpointType> {
                 multiInstance: database.multiInstance, // TODO is this always false?
                 options: {},
                 schema: metaInstanceSchema,
-                password: database.password
+                password: database.password,
+                devMode: overwritable.isDevMode()
             }),
             addConnectedStorageToCollection(
                 this.collection,