+++ src/plugins/utils/utils-rxdb-version.ts
@@ -1,0 +1,4 @@
+/**
+ * This file is replaced in the 'npm run build:version' script.
+ */
+export const RXDB_VERSION = '14.0.5';